requirements.txt
aiogram==2.23.1
aiohttp==3.8.1
PyYAML==6.0
vkbottle==4.3.12
vkbottle_types==5.131.146.13
